// src/types.ts
export type SearchResult = {
  lat: number;
  lon: number;
  name: string;
  region: string;
  country: string;
};

export type Location = SearchResult & {
  tz_id: string;
  localtime_epoch: number;
  localtime: string;
};

export type Weather = {
  location: Location;
  current: Current;
};

export type Current = {
  last_updated_epoch: number;
  last_updated: string;
  temp_c: number;
  temp_f: number;
  is_day: number;
  condition: Condition;
  wind_mph: number;
  wind_kph: number;
  wind_degree: number;
  wind_dir: string;
  pressure_mb: number;
  pressure_in: number;
  precip_mm: number;
  precip_in: number;
  humidity: number;
  cloud: number;
  feelslike_c: number;
  feelslike_f: number;
  vis_km: number;
  vis_miles: number;
  uv: number;
  gust_mph: number;
  gust_kph: number;
};

export type Condition = {
  text: string;
  icon: string;
  code: number;
};